MCR-2898 improvements mentioned in review
* added license header
* reuse code as much as possible
* avoid log messages with cound 0 if no data is inherited
